app/src/androidTest/java/com/utek/android/utekapp/UtekAppDatabaseTest.kt
package com.utek.android.utekapp

import androidx.room.Room
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.platform.app.InstrumentationRegistry
import com.utek.android.utekapp.database.UserInfo
import com.utek.android.utekapp.database.UserInfoDao
import com.utek.android.utekapp.database.UtekAppDatabase
import org.junit.After
import org.junit.Assert.assertEquals
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException

/**
 * This is not meant to be a full set of tests. For simplicity, most of your samples do not
 * include tests. However, when building the Room, it is helpful to make sure it works before
 * adding the UI.
 */

@RunWith(AndroidJUnit4::class)
class UtekAppDatabaseTest {

    private lateinit var userDao: UserInfoDao
    private lateinit var db: UtekAppDatabase

    @Before
    fun createDb() {
        val context = InstrumentationRegistry.getInstrumentation().targetContext
        // Using an in-memory database because the information stored here disappears when the
        // process is killed.
        db = Room.inMemoryDatabaseBuilder(context, UtekAppDatabase::class.java)
            // Allowing main thread queries, just for testing.
            .allowMainThreadQueries()
            .build()
        userDao = db.userInfoDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun insertAndGetNight() {
        val userInfo = UserInfo()
        userDao.insert(userInfo)
        val user = userDao.getUser()
        assertEquals(user?.sessionId, "init sessionId")
    }
}